requirements.txt
Django==1.11.11
djangorestframework==3.9.2
mysqlclient==1.4.2.post1
PyMySQL==0.9.3
pytz==2019.1
